Initial table for unsent devices follow-up
