+++ libs/src/configs/index.ts
@@ -1,4 +1,5 @@
 export * from './actionables-management-config';
+export * from './unsent-devices-management-config';
 export * from './discrepancy-management-config';
 export * from './order-logs-config';
 export * from './order-management-config';
