+++ libs/src/configs/unsent-devices-management-config.tsx
@@ -1,0 +1,46 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+import { isEmpty } from 'lodash';
+import { formatDate, parseStatus } from '../helpers';
+import { OrderItemStatus } from '../constants';
+
+interface ParsingFunctionParams {
+  row: { [key: string]: any };
+  menuItems?: any;
+  index: number;
+}
+
+export const unsentDevicesManagementParsingConfig = {
+  'Order ID': ({ row }: ParsingFunctionParams) => {
+    if (!row || isEmpty(row['order_number'])) return '--';
+    return row['order_number'];
+  },
+  'Order Date': ({ row }: ParsingFunctionParams) => {
+    if (!row || isEmpty(row['createdAt'])) return '--';
+    return formatDate(row['createdAt']);
+  },
+  'Order Type': ({ row }: ParsingFunctionParams) => {
+    if (!row || isEmpty(row['credit_type'])) return '--';
+    return parseStatus(row['credit_type']);
+  },
+  'Customer Name': ({ row }: ParsingFunctionParams) => {
+    const userDetails = row ? row['user_id'] : null;
+    if (
+      !userDetails ||
+      (isEmpty(userDetails['first_name']) && isEmpty(userDetails['last_name']))
+    )
+      return '--';
+    const firstName = userDetails['first_name'] || '';
+    const lastName = userDetails['last_name'] || '';
+    return `${firstName} ${lastName}`;
+  },
+  'Extended Prior': ({ row }: ParsingFunctionParams) => {
+    if (!row || isEmpty(row['isExtended'])) return 'No';
+    return row['isExtended'] ? 'Yes' : 'No';
+  },
+  'Devices Awaiting': ({ row }: ParsingFunctionParams) => {
+    if (!row || isEmpty(row['order_items'])) return '--';
+    const orderItems = row ? row['order_items'] : null;
+    const orderCount = orderItems.filter((item: any) => item.status === OrderItemStatus.CREATED)?.length;
+    return orderCount;
+  },
+};
