+++ libs/src/constants/index.ts
@@ -1,4 +1,6 @@
 export * from './constants';
+export * from './interfaces';
+export * from './table-columns';
 export * from './enums';
 export * from './interfaces';
 export * from './mock';
